+++ WickedEngine/WickedEngine_SOURCE.vcxitems
@@ -249,12 +249,11 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)Utility\tinyddsloader.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Utility\vk_mem_alloc.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)Utility\volk.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)wiAllocators.h" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)wiAllocator.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiArchive.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiAudio.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiAudio_BindLua.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiCanvas.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)wiContainers.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiECS.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiEvent.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)wiFFTGenerator.h" />
